Amended Hero component
